Explicitly cast tid to uint32_t

+++ OrbitBase/Tracing.cpp
@@ -85,7 +85,7 @@ void Stop() {
   auto& scope = GetThreadLocalScopes().back();
   scope.end = MonotonicTimestampNs();
   scope.depth = GetThreadLocalScopes().size() - 1;
-  scope.tid = orbit_base::GetCurrentThreadId();
+  scope.tid = static_cast<uint32_t>(orbit_base::GetCurrentThreadId());
   TracingListener::DeferScopeProcessing(scope);
   GetThreadLocalScopes().pop_back();
 }
@@ -94,7 +94,7 @@ void StartAsync(const char* name, uint64_t id, orbit::Color color) {
   TracingScope scope(orbit_api::kScopeStartAsync, name, id, color);
   scope.begin = MonotonicTimestampNs();
   scope.end = scope.begin;
-  scope.tid = orbit_base::GetCurrentThreadId();
+  scope.tid = static_cast<uint32_t>(orbit_base::GetCurrentThreadId());
   TracingListener::DeferScopeProcessing(scope);
 }
 
@@ -102,7 +102,7 @@ void StopAsync(uint64_t id) {
   TracingScope scope(orbit_api::kScopeStopAsync, /*name*/ nullptr, id);
   scope.begin = MonotonicTimestampNs();
   scope.end = scope.begin;
-  scope.tid = orbit_base::GetCurrentThreadId();
+  scope.tid = static_cast<uint32_t>(orbit_base::GetCurrentThreadId());
   TracingListener::DeferScopeProcessing(scope);
 }
 
@@ -123,7 +123,7 @@ void AsyncString(const char* str, uint64_t id, orbit::Color color) {
 void TrackValue(orbit_api::EventType type, const char* name, uint64_t value, orbit::Color color) {
   TracingScope scope(type, name, value, color);
   scope.begin = MonotonicTimestampNs();
-  scope.tid = orbit_base::GetCurrentThreadId();
+  scope.tid = static_cast<uint32_t>(orbit_base::GetCurrentThreadId());
   TracingListener::DeferScopeProcessing(scope);
 }
